feat: add check multisig, add stages callback
