+++ packages/core/src/common/utils/index.ts
@@ -1,0 +1,2 @@
+export { checkIsContract } from "./checkIsContract";
+export { getFeeData, FeeData } from "./getFeeData";
