+++ packages/stake/src/common/constants.ts
