+++ packages/stake/src/types.ts
@@ -1,3 +1,4 @@
+import { type Address } from "viem";
 import {
   type LidoSDKCoreProps,
   type LidoSDKCore,
@@ -5,4 +6,22 @@ import {
 
 export type LidoSDKStakeProps = LidoSDKCoreProps & {
   core?: LidoSDKCore;
+};
+
+export enum StakeCallbackStage {
+  "SIGN" = "sign",
+  "RECEIPT" = "receipt",
+  "CONFIRMATION" = "confirmation",
+  "DONE" = "done",
+  "MULTISIG_DONE" = "multisig_done",
+}
+
+export type StageCallback = (
+  stage: StakeCallbackStage,
+  payload?: unknown
+) => void;
+export type StakeProps = {
+  value: string;
+  callback: StageCallback;
+  referralAddress?: Address;
 };
